Work around live streams
As of JojPlayer 5.4, previously added live streams do not work anymore.
In JojPlayer 5.4 there has been added 'auth' parameter in a rtmp stream
play path which cannot be omitted. Value of 'auth' parameter is MD5 of
'authentication' string returned by the server and salt, which is
'malekoniky'. The 'authentication' string consists of UTC time returned
by the server and a random number. The string looks like this:
'2015/03/08 22:30:32.714 56626627'. Unfortunately, it seems like there
is no other way to obtain such a string without using JojPlayer. Now,
the plugin uses an API designed for JOJ's mobile application. It does
not support HD streaming and Senzi TV.